+++ src/Curiosity.Migrations.PostgreSQL/Properties/AssemblyInfo.cs
@@ -2,4 +2,4 @@
 
 [assembly: InternalsVisibleTo("Curiosity.Migrations.PostgreSQL.UnitTests")]
 [assembly: InternalsVisibleTo("Curiosity.Migrations.PostgreSQL.IntegrationTests")]
-[assembly: InternalsVisibleTo("Curiosity.Migrations.TransactionTests")]
+[assembly: InternalsVisibleTo("Curiosity.Migrations.IntegrationTests")]