+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/DependenciesTests/DependencyCodeMigrations/IDependencyMigration.cs
@@ -1,0 +1,6 @@
+﻿namespace Curiosity.Migrations.IntegrationTests.DependenciesTests.DependencyCodeMigrations;
+
+public interface IDependencyMigration
+{
+    
+}