+++ tests/IntegrationTests/Curiosity.Migrations.IntegrationTests/TransactionsTests/TransactionCodeMigrations/ITransactionMigration.cs
@@ -1,0 +1,6 @@
+﻿namespace Curiosity.Migrations.IntegrationTests.TransactionsTests.TransactionCodeMigrations;
+
+public interface ITransactionMigration
+{
+    
+}